prediction_api/requirements.txt
numpy==1.19.2
scikit-learn==0.24.2
tensorflow==2.5.0
flask==2.0.1
python-dotenv==0.17.1
gunicorn==20.1.0
